Add GalleryState class and GalleryImage data class, refactor RequestState
