+++ app/src/main/java/com/xdiach/diarymoodapp/data/repository/MongoRepository.kt
@@ -1,7 +1,7 @@
 package com.xdiach.diarymoodapp.data.repository
 
 import com.xdiach.diarymoodapp.model.Diary
-import com.xdiach.diarymoodapp.util.RequestState
+import com.xdiach.diarymoodapp.model.RequestState
 import java.time.LocalDate
 import kotlinx.coroutines.flow.Flow
 import org.mongodb.kbson.ObjectId
