+++ app/src/main/java/com/xdiach/diarymoodapp/model/GalleryState.kt
@@ -1,0 +1,34 @@
+package com.xdiach.diarymoodapp.model
+
+import android.net.Uri
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.remember
+
+@Composable
+fun rememberGalleryState(): GalleryState {
+    return remember { GalleryState() }
+}
+
+class GalleryState {
+    val images = mutableStateListOf<GalleryImage>()
+    val imagesToBeDeleted = mutableStateListOf<GalleryImage>()
+
+    fun addImage(galleryImage: GalleryImage) {
+        images.add(galleryImage)
+    }
+
+    fun removeImage(galleryImage: GalleryImage) {
+        images.remove(galleryImage)
+        imagesToBeDeleted.add(galleryImage)
+    }
+
+    fun clearImagesToBeDeleted() {
+        imagesToBeDeleted.clear()
+    }
+}
+
+data class GalleryImage(
+    val image: Uri,
+    val remoteImagePath: String = ""
+)
